fixed a bug in the buttonremapping script
previously if this script was on a button it would not display the correct text caption when retrieving from a save file. I fixed this by allowing some of the mapped controls to set the keycode to none and instead use a string array to chance the caption on the button.